requirements.txt
cached-property==2.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
dnspython==2.7.0
exchangelib==5.5.1
idna==3.10
isodate==0.7.2
lxml==5.3.1
oauthlib==3.2.2
pycparser==2.22
Pygments==2.19.1
pyspnego==0.11.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
